added tutorial scene
also a credits scene but it's currently non-functional. Currently the
game starts on the tutorial page, but I temporarily disabled the "fade
to black and start the level" functionality so our designers could work
on the fonts. In order to test the gameplay, uncomment two lines in
/scripts/tutorial.gd - specifically the two that set "fading_out" and
"fading_in" to true and false. Just remove the # symbol at the start of
these two lines and you should be good.