fixed flashing main menu bug
when moving from the main menu to the credits scene or the tutorial
scene, the screen would flash with the contents of the new scene before
dropping to black and slowly fading in. It was a little jarring. This
commit should fix that functionality such that when the new scene is
loaded, the image starts from black and fades in correctly.